[zerror] Require errors implement a trait.
The wrapper is nice, but ugly in type signatures.  It also isn't very
extensible.

Introducing Z.  Z is a trait to be implemented for dealing with errors.
To complement Z is ErrorCore that implements meaningful versions of Z,
so that all a new type must do is wrap ErrorCore and proxy Z.  But Z can
be extended if need be.