[guacamole]  Fix the zipf distribution to be skewed, not essentially uniform
When implementing zipf I copied the Jim Gray paper, but there was an
argument swap in zeta2, which leads to it being a not-Zipf distribution.

Because the alpha and theta parameters don't give proper distributions
of zipf (as it's only an approximation), I've switched to a "skew"
parameter that goes (0.0, 1.0) exclusive.